type changes + bullet work + BuildMaster
- changed speed to be s8 instead of s16
- broke up the bullet update to smaller functions that handle checking against one edge

BuildMaster
- re-added max-allocs-per-node i

+++ project/source/client/ai_commands.c
@@ -29,7 +29,7 @@ void AICommandUtils_updateAICommandsRunner(struct game_object* object, AICommand
 			aiCommandsRunner->currentAICommandItem = (const AICommandItem*)data;
 			break;
 		case AI_COMMAND_SET_SPEEDX:
-			object->speedx = (s16)data;
+			object->speedx = (s8)data;
 			aiCommandsRunner->waitTime = 1;
 			return;
 		case AI_COMMAND_SET_SPEEDY:
